Move store/notifications to utils/
